The notebook text is now entered
The notebook text is now entered, but needs some further polish.

Here is also a small difference in the result when noise is introduced compared to the previous  version in BPL_light. The deterministic case is exact the same.

I have checked: sigma, sample intervall.
Perhaps something with the seed....